Telegram profile pictures and display names are updated
The bridge updates the displayname and/or the avatar of the bridged Matrix
user when the Telegram user changes them.

This is done by checking the name and profile pic of the sender when a message
is received from Telegram, and comparing to the database to see if they have
changed. If they have, it will update the database and send the updates to the
Matrix homeserver.